+++ ThirdPartyAdapters/facebook/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Facebook Adapter for Google Mobile Ads SDK for Android Changelog
 
+## 5.2.0.1
+- Updated adapter to support new open-beta Rewarded API.
+- Updated the minimum required Google Mobile Ads SDK version to 17.2.0.
+
 ## 5.2.0.0
 - Verified compatibility with Facebook SDK v5.2.0.